requirements.txt
apache-libcloud
configparser
pycrypto
requests
typing
